test: isolate parameter extractor model fixtures
